requirements.txt
discord.py>=2.3.2
playwright>=1.45.0
aiohttp>=3.9.0
openai>=1.45.0
pydantic>=2.4.2
python-dotenv>=1.0.1
tenacity>=8.2.3
beautifulsoup4>=4.12.3
playwright-stealth>=1.0.6